emotion/generic - Set filename to NULL on file_close.
Otherwise if we try to call file_set(<some file>) after file_set(NULL),
it will send a close command to the player for an already closed file.

Also make the example cycle through the list of movies, but still
passing through a NULL file when reaching one of the ends of the list.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/emotion@63208 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33