emotion/generic: Fix frame dropping.
Move the frame drop counter to the shared memory object, so the player
can check its value and not send new frames if we didn't "consume" the
previous one.

Also move the triple buffering pointer changing to the display function,
so it will only lock once when accessing the critical region.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/emotion@63220 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33