emotion/generic/vlc: VLC needs to write data somewhere.
It seems that depending on the system, vlc can't use a NULL pointer to
the pixels where it should write its data.

So a small amount of memory should be allocated and passed to its
rendering callbacks (specifically, the lock callback) when the file is
being opened and decoded for the first time. Then this memory can be
freed, since the real rendering will happen on the shared memory area.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/emotion@63777 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33