emotion: make Xv work.
NOTE: we need the help of the window manager to make this
really work. So for the moment, it half work. As soon as I
hack E17, the Xv fast path for Emotion will only work when
used with E17 and it will fallback to canvas inlined rendering
in other case.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/emotion@63802 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33